Update README.md
Customer relationship management (CRM) is software which is created to assist your business in fostering better relationships with your customers and provide them with a seamless experience.

Without a sound CRM strategy, your customer data will be scattered all over the place, making it difficult to keep track of all the contacts and touch points you've had with specific clients. As a result, the sales cycle will get delayed and the client experience will be flawed.

Why does your business need a CRM Software?

Lack of centralization:

When your client information is scattered across Excel sheets, business cards, and meeting notes, it's challenging to make significant decisions. The process of collecting and making meaning of the data is time- and effort-consuming. 

Customer service is compromised:

Getting clients is already a challenge and losing them is the biggest flaw. It may be time to invest in a CRM system if you discover that your customer service team spends more time passively responding to customer complaints than proactive anticipating customer requirements.

Unable to keep up with lead flow:
CRM may be necessary if you're in the fortunate situation of experiencing a flood of new leads but are finding it difficult to keep up with it. Your leads will be quickly sorted and prioritised by an intelligent CRM system, enabling your sales team to react to them in a prompt and effective manner.

When choosing the CRM tool for your business, the first choice you'll be making is between on-premise or cloud CRM software. Choosing an on-premise CRM software like OneHash will let you keep track of existing leads and deals in your sales funnel and pipeline, integrate with omnichannel media platforms and explore which platforms yield the most results, analyse why, where, and how you lost a deal.